requirements.txt
addict==2.4.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
filelock==3.4.1
gitdb==4.0.9
GitPython==3.1.18
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
MarkupSafe==2.0.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
mmcv==1.6.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1621287993989/work
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
sentry-sdk==1.9.8
setproctitle==1.2.3
shortuuid==1.0.9
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0.post1
smmap==5.0.0
soupsieve==2.3.2.post1
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
timm==0.6.7
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
yapf==0.32.0
zipp==3.6.0
